standalone_agent/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
python-docx==1.1.0
google-generativeai==0.7.2
graphviz==0.20.3
Pillow==10.3.0
pydantic==2.8.2
mistletoe==1.3.1
